renamed clerks to views
