+++ views/Dot/Dot.java
